requirements.txt
py-cord
PingPongWr
firebase_admin
python-dotenv